web_app/requirements.txt
gspread 
google-auth 
requests 
python-dotenv
uvicorn 
selenium
fastapi
gunicorn
pydantic